centrallog: add first prototype of webrequest-live with Benthos
The change adds an instance of Benthos for each centrallog node,
sending 1/128th of the traffic pulled from the webrequest_{text,upload}
Kafka topics to a new one called webrequest_128_live. The data
is enriched with Geo-location to mimic what the Data Engineering's
batch job do regularly after importing webrequest data.

The long term idea is to use k8s for Benthos, but as first step
for this use case the centrallog nodes seem a good fit. Due to how
the Kafka consumer groups work, we'll have two active instances
on both centrallog nodes handling half of the Kafka partitions.
If one of the nodes go down, the other one will pick up its partitions
transparently. Not a bullet-proof design but it should guarantee to us
a minimal redundancy in case of failures/maintenance/etc..

Bug: T314981
Change-Id: I90ec3862d483e7981e1f74f93f48bfdca0f8c963