Configure the kube_env file for the spark-operator namespace
We will be deploying spark-operator using standard helmfile mechanism,
which means that we need to configure a kube_env file containing the
tokens. We are not doing so for the spark namespace at the moment, since
we believe that this will use an alternative mechanism.

The spark-operator and spark-operator-deploy tokens are already in place
in the private repo.

Bug: T321686
Change-Id: I95d7fe7dbf5322ee9ebaae8e40e7eeba7b06d373