homer: Don't accept a commit with an empty message
Normally if the message is empty, git aborts the commit, so users are
used to that behavior. The bug here is that we prepend (username) to the
commit message, and then it isn't empty anymore. This fix duplicates the
logic used at modules/puppetmaster/files/git/private/commit-msg-master,
where we had the same problem.

Change-Id: I18cd4d5c3c37378cde8b48e98b8f9c332dd8df57