dumps/distribution: fix values that don't fit into data types
As part of I4cdcc1b67873f937 we are trying to add a new custom
data type, Wmflib::Dumps::Mirror, which describes the data structure
for a dumps mirror.

It expects the "active" field to be either yes or not but not something
else (Stdlib::Yes_no), so I am replacing strings like "notyet" with a
simple "no".

Also setting some strings where there are currently none.

This should solve the errors on:

https://puppet-compiler.wmflabs.org/pcc-worker1003/38057/clouddumps1002.wikimedia.org/change.clouddumps1002.wikimedia.org.err

Change-Id: I77b234e0a889ddb40d27e83fcc75739206e88457