cloudgw: introduce more robust vlan interface naming
Introduce support for vlan interface naming on the form `vlanXYZW` instead of
the old `nic.XYZW` which can be problematic if the base NIC has a long name (which is
the case for cloudgw1001/1002).

A hiera key is introduced to keep old behavior while servers are reimaged in the new
setup.

The first host to be migrated to the new setup is cloudgw2002-dev, and the rest will
follow soon.

Bug: T319184
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: Ib274249b0935a7a818e28fc7096827f3de712234