snapshot: Parallelize cirrus dumps by db shard
CirrusSearch changed over to use a more resilient, but slower, method
of generating the dumps in Ifb5f043e3d. To get the dumps to complete
in a reasonable time frame parallelize the dumps by running the script
once for each database shard.

As part of this the dump of CirrusSearch's metastore index has been
removed. With the sharding it would have been repeatedly dumped for
each shard, fixing would mean another timer/script to only dump this
index. The index itself has been less used over time, it's rarely
referenced in production other than for state of a process that loops
through the other indices checking for out-of-date revisions.

Bug: T265056
Hosts: snapshot1008.eqiad.wmnet
Change-Id: I311f99a0b95517c568a63256ef6afa19de867e8e