database-backups: Update partman recipe for dbprov1004 / dbprov2004
We use the regular db formatting recipe to install dbprov hosts,
with the difference that sda should use HW RAID 6, rather than
RAID 10 on dbs, and that dbprovs should have an additional 2
SSD hosts in RAID 0 (also in hw).

The SSDs are setup post-install as they are used for scratch data
and mounted on a specific directory created by puppet.

Bug: T322256
Change-Id: Ib95f38052b88c5c78f8100a40d105567d7addd1a