Patch cinder volume_type api to allow non-uuid project ids.
Uniquely among the entire OpenStack ecosystem, the cinder
'volume type' APIs validate the project ID and reject
it if it isn't a UUID. This is /technically correct/ but
breaks our use of project_id==project_name convention.

This causes the 'openstack volume type set --project' command
to fail.

This is upstream patch
https://review.opendev.org/c/openstack/cinder/+/864585

Bug: T301949
Change-Id: Iaeb338904f1f2f3107e76dafc60a4068d4745fae