phabricator: use systemd::sysuser for phd user, also on phab1004
This is already used on phab2002 but not on phab1004.

It means the 'phd' system user for Phabricator gets managed
by systemd sysuser.

This in turn means it can have the globally reserved UID/GID
and finally has the same UID and GID and we should never again
run into privilege issues when rsyncing data between Phabricator hosts.

phab2002 and phab1004 should be the same though while phab1001 is NOT changed
and phab2001 is gone.

Bug: T280597
Change-Id: I4498e533feb53875de04f1486f60a527a98552d4