gitlab_runner: make one Shared Runner canary
This change adds a canary tag to runner-1030 in gitlab-runners
wmcs project. The tag makes it easier to test changes and run dedicated
test jobs on the runner.

If tags: [canary] is used in a .gitlab-ci.yml job, the job will
get scheduled on this specific runner. This change will not prevent
the Runner from accepting normal, non-canary jobs. run_untagged
is still set to true.

Change-Id: I92a0532787e2dac8d86c721d03db76df79836f14