cloudvirts: introduce modern NIC setup and use it by default
This is controlled by a new hiera key, which implies:
 * use a single NIC for dataplane / controlplane traffic
 * make use of the `vlanXYZW` naming scheme
 * is the new default behaviour

This is to prevent problems with too long NIC names in the old scheme `eno1.XYZW` and
to eventually simplify the whole setup (once all cloudvirts are in the new scheme).

My previous approach of just using "interface_primary.XYZW" wont work in eqiad1
where several cloudvirts have long interface names.

The intended workflow would be to reimage each cloudvirt and merge a patch
removing unused host-level hiera to enable the single NIC setup.

This patch also moves cloudvirt2001-dev to the new setup. It will be reimaged.
Host-level hiera is introduced for 2002/2003 so they don't change until they
are reimaged.

Bug: T319184
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: I9a5de108f7cd535b87c92f9beb070025c5bbc85c