benthos: reduce webrequest-live kafka partitions to read
We know that varnishkafka produces messages/events to
the webrequest topic partitions using a random distribution,
so it makes sense to reduce the amount of data read as input
by Benthos (and hence processed) since the sampled data
should be equally representative. This should allow us to
reduce the consumption of network bandwidth and load on
centrallog nodes.

The configuration has been adapted to set the sampling
value from hiera, since reducing the partitions means also
changing the sampling value (if we want to keep the same 1/128th
ratio).

Bug: T314981
Change-Id: I1a6c6fe3cd61822e588941553589968c2baac2b5