phabricator: remove hardcoded ports, use parameters in my.cnf for admins
Avoid having the hardcoded port number in this template.

Use the existing class parameters to set the mysql_slave
host name and port instead.

Remove the use of $domain because the FQDN is in Hiera per DC
anyways.

To avoid having to add the parameters to the main phabricator class
and because it probably belongs there anyways, move this code
from the main module to the profile.

Bug: T280597
Change-Id: I9ea2a6d9a1e04e503f0a2de55c9b3cea5bf106be