profile::kubernetes::node: add k8s_116 tag to infrapod's default
The current docker image pulled for the pause container should be
"latest", since we have been having only one value in the registry
up to now. We are going to publish a new image soon, that will take
the "latest" tag becoming the default across clusters.

We pushed a new tag called "k8s_116", already tested in staging,
so let's change the hiera's default before publishing the new
pause image to the registry.

Hosts: auto
Bug: T322920
Change-Id: Iccf650ea9f7ad1ab8dbe48e4a10af4b87cd32825