phabricator: set mysql master port for eqiad
Currently we are setting these for role/codfw but not for role/eqiad.

The host names change between eqiad and codfw, the ports don't right now
but might at some time?

Even if the port could technically be in role/common I guess
I prefer having it with the host name here and be explicit.

Bug: T280597
Change-Id: I074419b89984677b5716c9efc72ccce153b0cbd5