Add logind to filter list for debdeploy
Don't flag necessary restarts for systemd-logind. logind restarts
are very intrusive since they log out users. There's various
libraries linked against logind, but there's very little attack
footprint, most would entail some level of control of configuration
which poses a risk by itself. If we're ever in a situation where
an upgrade of an underlying library warrants a restart of logind
we'd orchestrate it via Cumin manually in a way that checks for
impacted users (or announce a maintance window or reboots).

Bug: T135991
Change-Id: Iac2ed9aea00821b6b2d3160973a421b804dbefdb