mjolnir msearch: Reduce allowed concurrency
Mjolnir was initially designed with the concept that there would be an
idle cluster it could throw queries at. With mediawiki now serving
read requests from secondary datacenters that is no longer the case
and this mseearch daemon has been sitting idle as it could not find
an unloaded cluster to submit queries to.

To bring it back into operation reduce the allowed concurrency
significantly, such that the daemon can only consume a fraction of
the cluster resources that should be available at any time from
any cluster.

The primary change is reducing the number of workers per dc from 8 to
2. The value for max_concurrent_searches was also reduced from 2 to 1.
In current usage none of the queries sent to this daemon utilize
concurrent searches, reducing to 1 should keep the daemon from
increasing cluster utilization if some future usage does.

Bug: T318575
Change-Id: I0b262c27ee2d29d7e9668dfcd9fafdb22360f7b8