mediabackups: Add new policy intended for admin deletion of files
Files can be searched and restored with the minio (S3 API) user
for automated restores. However, it cannot delete files.

As automation right now uses the same account, add a new policy
to list, read (download) and delete files, and apply it to
the restore user.

In the future we may want to have a separate account just to
allow listing and deletion, but this will unblock for now
deletions without needing to affect admin procedure.

Bug: T323796
Change-Id: I2e7d519d93bc96c328ddb72d486aeb15504aed44