docker_registry_ha: remove unused cache::nodes ref
Looks like the purpose of this was gutted over a year ago in
349ab224 when we got rid of traffic edge nodes connecting to this
over plain HTTP on port 81.  This eliminates the last reference to
cache::nodes in the puppet repo outside of traffic's "internal"
use for backend chashing.

The derived variable $http_allowed_hosts seemed to be unused as
well, so I took the liberty of removing it also.

Bug: T256762
Change-Id: I2e4d8b2182210925d13d34c12af309dc1a4a2a25