base: fix puppet_alert.py
Fixes a check for catalog compilation failures in puppet_alert.py.
Previously it checked if the current run had exactly zero changes, but
it should have checked if the compiled catalog has exactly zero
resources. This only started being a problem after [0] was merged, since
the notify{} resource it removed counted as a change.

[0]: I212fa3a4b2311a7695de2e38e660a3e969b866d5

Change-Id: I2a31e8fb2ac081bef2e52c79a7d095dc0b50faf3