P:cache::haproxy: harden systemd unit
The haproxy systemd unit utilized by the cp hosts should be hardened,
and this commit is a first step towards that. The plan is to work on
incrementally hardening the unit by first starting with the "obvious"
sandboxing options and then restricting the service capabilities even
further.

Bug: T323944
Hosts: cp4045.ulsfo.wmnet
Change-Id: I4f25608ba9e3f7b4b796c4b48d4ede146e7b9d72