Openstack config: move keystone_authtoken into a shared template
This replaces www_authenticate_url with www_authenticate_uri in
many places but oddly www_authenticate_url doesn't seem to be
documented in modern OpenStack anyway.

This also moves the internal and public keystone ports
out of a hiera lookup and into a template. This is a bit
less flexible but it greatly simplifies the manifests
so I think it's worth it.

Change-Id: Ie867100b0f6843e48a2fc429e35fbf511d2e203f