Use parquet column names to order results from hive in presto
This settings changes the behaviour of presto when returning results
from parquet. When this setting is false, which is the default, the
ordinal position of the column is used to determine the result.

When true, the results are ordered according to the names of the columns
which is desirable, especially when schema evolution has occurred.

Bug: T321960
Bug: T321231
Change-Id: Ic507540685114548b250b87b5bbbd17ccc0be30b