gerrit: raise H2 compaction time
For caches backed up by a disk, Gerrit uses the H2 Database. It is also
used to keep track of files that have been reviewed in
`account_patch_reviews.h2.db`.

When doing the Gerrit 3.4 > 3.5 upgrade, the online reindexing of all
changes caused some entries to be added to two of the largest caches
`gerrit_file_diff` (8G) and `git_file_diff` (11G). That has lead to the
disk (40G) filling up entirely and turned the maintenance window to an
outage.

The disk size reported by Gerrit were respectively 157M and 158M which
has lead me to believe the files were not vacuumed to reclaim freed data
blocks. A gzip of one of the file yields a 32MBytes archive, a good
indication the large files are mostly empty.

We have most probably been carrying out the caches from a server to
another one since we first installed Gerrit (2010?). The H2 databases
are compacted upon disconnection and only allows a 200 milliseconds
window for that. That is not enough.

We can allocate 15 seconds (15 000 milliseconds) to the driver by
passing to the jdbc URL `;MAX_COMPACT_TIME`. The URL is in the Gerrit
java code and not configurable though. I found out H2 looks up settings
down to system properties in camel case. Thus the equivalent of
`MAX_COMPACT_TIME` is the property `h2.maxCompactTime`.

To activate the setting, the Gerrit JVM will have to be restarted. Upon
the next restart, H2 should then further compact the files.

NOTE: Gerrit currently takes a long time to shutdown and ends up being
SIGKILL by systemd after 90 seconds. It is a different problem though.

Bug: T323754
Hosts: O:gerrit
Change-Id: I0b141bcbf509d8e5e0614d14d3b817afbe25a0a1