contint: add ci::master to contint1002
This is the preliminary Puppet provisionning to bring up contint1002 has
a secondary CI system. Ultimately it will run the zuul-merger and be
attached as a Jenkins agent (that is done manually).

None of Jenkins, Zuul scheduler or the Zuul merger should run at this
point. The services can be disabled via hiera data passed to the
profiles which I have applied in contint1002.yaml. The role::ci::master
defaults for each three services being:

- Jenkins is masked/stopped
- Zuul merger is running
- Zuul server is running/enabled

Add the new host to various access lists which thus also affect the
current primary (contint2001) and contint1001.

The new host contint1002 is intentionally not added to the scap dsh
group yet since it needs this change to provision it as a scap target.

Bug: T313832
Hosts: contint1001.wikimedia.org
Hosts: contint1002.wikimedia.org
Hosts: contint2001.wikimedia.org
Change-Id: I3d6568404b7d62f60bf161d12f832abfe5ac07f1