ci: move docker::settings to common, avoid host names
When adding the production role on new host contint1002
we first had an error because the docker::settings could not
be found and needed a follow-up to add them to fix puppet.

Since they seem to be always the same, let's move them to common
and avoid settings based on host names.

If something is not common the next step would be to do it per DC.

Bug: T313832
Change-Id: I969edfcf1e5b5fdc88d1435d7e633dabbdad9edd