ci::master: hack to bootstrap new server contint1002
There is currently a puppet error when the ci::master role
is applied on a brandnew host, like contint1002.

It seems to have been introduced by Ie1b87c9f896bcb0
which expects an admin group, contint-admins, to exist
to be used in helm-related classes, before the admin
group has actually been created.

As a hack to get around this and bootstrap contint1002 for now,
since contint1001 is breaking, just comment it out for a moment.

Then will disable puppet on all contint*, run puppet only on
contint1002 and then revert this change.

This should get around the problem and a long-term fix can be
found separately.

Bug: T313832
Change-Id: I1dd23fb80f8a025955301607eba87e2167be3286