certspotter: temporarily disable certspotter (and the systemd timer)
As discussed in T318911, there are various issues with certspotter that
prevent it from being useful. Pending those issues being fixed,
certspotter doesn't provide any meaningful data that is being currently
reviewed because of the low signal-to-noise ratio.

As per discussions in the Traffic meeting, we are disabling certspotter
till we can improve it or fix the issues mentioned in T318911.

Bug: T318911
Hosts: O:alerting_host
Change-Id: Ib45e0115ae9050c417f5297c35c864a0d44f889b