puppetmasters: cache cleanup
We could potentially purge all old files in /var/lib/puppet
but this has a lighter touch and only deletes things that
we definitely don't need.

This will save a small bit of space in prod, and a ton
of space in cloud-vps where we churn through new VMs
every few minutes.

Change-Id: I8d8d0e9f3f4588187e7c678503e1bb5eab12bd34