Backing up HDFS FSImage to HDFS on Monday morning
The HDFS FSImage is sent to HDFS in order to run analysis on it. The
management of those HDFS files will be managed by the Airflow pipeline.
(analytics/dags/hdfs_usage/hdfs_usage_weekly_dag.py)

The target HDFS permission is `analytics:analytics-admins 750` in order
to keep this dataset private between HDFS users.

* Send the backup to HDFS on Monday at 2am
* Also compress the fsimage to save network & disk

Bug: T324850
Change-Id: Ia6796aec0d0f04d99f44a3e71788b59b480ad11b