posgresql: properly confine fact
A fact `confine` method is expected to return a boolean, the
`postgres_replica_initialized` fact only confined on `sudo` which was
the last call. Although presence of `psql` was checked, it did not
participate in the return value due to lack of a boolean AND. We need
both conditions to be set.

Bug: T324571
Change-Id: I77caed7bdd67f6cb21f7480a3f0e3ea9901d9221