Add two new cache configuration parameters for superset 1.5.2
Superset introduced two new cache types which should be specified from
version 1.5.0 onwards. Omitting them results in a performance penalty,
since it will default to the database backend.

We configure the superset-next instance to use memcached on localhost,
which is the same as the metadata cache.

Various small linting changes and type definitions have been
incorporated, but they are intended to be noops.

Hosts: an-tool1005.eqiad.wmnet,an-tool1010.eqiad.wmnet
Bug: T323458
Change-Id: I59f342793c38b80cc7f516fcfc342d4567e40d20