grafana: Explicitly set default theme to light
Why:
Let's actually be explicit about what the default theme we want is. We
've been using light theme for a very long time, but never really coded
it down to our puppet config.

It should be noted that both organization as well as individual users
can override the theme to one of their own choosing. The Wikimedia
organization does not override it currently (which is probably a
good thing)

What:

Set default_theme: light in grafana.ini

Change-Id: I902511be138fe6559f39ac6e533437210533d96b