site: add contint2002 to ci::master role
Apply the ci::master role on new host contint2002, along with other ci
hosts. As the comment suggests the actual switch to who is the master
among masters should be in Hiera, not here.

Ensure Jenkins, the Zuul scheduler and the Zuul mergers are all stopped
while the host is being completed.

Bug: T324659
Hosts: contint2002.wikimedia.org
Signed-off-by: Antoine Musso <hashar@free.fr>
Change-Id: I1992ed2c2c853d2d06cddc6ecc9b070c58067f9d