cloud: allow VMs to connect to contint1002 and contint2002
This looks like it is what allows VMs in Wikimedia Cloud
to connect to the CI master in production.

It would be ideal if we could avoid hardcoding host names / IPs
here and just get the data from the central hiera data.

But until then, we should just add the new host contint2002 that
is going to replace contint2001.

And to make sure everything works let's first add it in parallel
and later remove contint2001.

Additionally, add contint1002 because it replaced contint1001
just recently and if we ever need or want to fail-over then
we should reduce the number if places we have to edit.

Bug: T313832
Bug: T324659
Change-Id: Ia240db699933a40a5ed08938d9e45b1903fe151a