ci: add contint2002 to zuul_merger firewall, ferm_srange
Another step in preparation to getting contint2002 into production.

This, as other additions to firewall rules, should be possible
before other steps so we get things out of the way.

Overall I think it would be nice if we could reduce the number
of places where we list the same hosts, but that is separate.

Bug: T324659
Change-Id: I8df6c2ba17c31f484c0929b0f41eeafa2164c1ca