elasticsearch/relforge: add contint2002 to cirrus::ferm_srange
This server is going to replace existing contint2001.

Adding this to this firewall rule.

Fine any time before the host is actually in production, I assume.

I also think the number of places where we list CI server host names or IPs
is too high, but that's for another day.

Bug: T324659
Change-Id: I818790293c350276712271a9e6dda32cab09f358