Add vendored module bodgit/puppet-postfix
This commit adds a vendored module:

  https://github.com/bodgit/puppet-postfix

Along with its dependencies. This module and its dependencies are
licensed under the Apache-2.0 license.

Three additional patches on top of upstream have been added, which are
open pull requests:

  1. Add Debian Bullseye(11) support
  2. Fix support for /etc/aliases
  3. Debian: don't install Augeas lens for unix-dgram

This module has been used to build a testing configuration for replacing
our existing use of Exim. Thus far it seems to be a well built module
and is worth vendoring based on my review of the code.

Dependencies:

1.  `bodgit/bodgitlib`:

    Unneeded currently, used for ldap, memcache, mysql, & pgsql lookups,
    we could always add it if we started using those postfix lookup
    types.

2.  `herculesteam/augeasproviders_core`:

    Alternative Augeas provider module which is now under Voxpupuli,
    this will be updated to their version once a new version is cut
    which includes my recent pull request,
    <https://github.com/voxpupuli/puppet-augeasproviders_core/pull/42>

3.  `camptocamp/augeas`:

    Unneeded, as we already have the updated postfix lens

Bug: T325396
Change-Id: I1f50481e6f06ba802c90c7b94057eacb71c10582