mariadb: Setup new definitive databases for mediabackups & bacula
mediabackupstemp is going to be removed. Before then, we have
to setup the new definitive hosts to host them- 2 sections called:

* backup1-eqiad
* backup1-codfw

The reasoning is as follows:
mediabackups is no longer suitable, as they will contain both
mediabackups data & bacula data. This section is special in that
it will have redundant but different data on each datacenter in
case a dc becomes corrupted or unavailable.

These are misc hosts, but marking them with the backup name,
to separate it from regular misc (m) hosts, that
have the same data on both dcs. 1 to leave space for other
(at the moment, not planed) backup-related sections with the same
setup.

Bug: T313582
Change-Id: If538a67cdf0b813b38345d73dd9aecb69f9ed9b3