Increase max_connections on analytics_meta MariaDB
We need to increase this value to give us more headroom, since this
database now hosts several airflow instances, druid, superset, datahub,
plus the hive metastore.

Note that Airflow databases will be migrated to PostgreSQL shortly.

Also the analytics_meta database is to be moved to dedicated hosts named
an-mariadb100[1-2] in the near future as well.

The configuation will be revewed again during and after these operations.

Bug: T325278
Change-Id: I77630edd0614fed4069901c6503dc7255b1cf07d