swift: move accounts_keys to common hiera global_account_keys
We need to be able to access both mw_media keys from the same host
(for swiftrepl replacement), which means that these need to be in
common heira. The rename makes the transition easier (since we can
make the private puppet change that introduces the new hiera item;
then deploy this; then remove the old accounts_keys hiera item.

This needs co-ordinating with
https://gerrit.wikimedia.org/r/c/labs/private/+/868718

And the equivalent change in actually-private puppet

Bug: T162123
Hosts: auto
Change-Id: I3b1164e7d2460a6387d680905a8eeb27642c0aab