deployment_server: add keyholder/group config for jenkins-ci deploy
After modules/secret/secrets/keyholder/deploy_jenkins has been created
and Ia5dde9cafbfa384 was merged, now we need the new key to show up in
/etc/keyholder.d on deployment servers.

This should do it.

But we need to pick one trusted group and there are several ones.

We are creating a new group, deployment-jenkins, in If17b910e0d86924
for this.

Bug: T324014
Change-Id: I111d8c9e2c0da5151f46cb49f84e27d2b2c741b4