Add another 12 GB of RAM to the presto server JVM
We have observed excessive garbage collection when increasing the size
of the presto cluster from 5 to 15 nodes. Therefore we believe that it
would be prudent to increase the amount of RAM available to the presto
coordinator.

The coordinator has at least 19 GB of unused (neither for buffer nor O/S
cache) RAM, so we feel that this is safe to do.

Bug: T325331
Change-Id: Ic03c74f268cbddb07443746b55518e83775dc201