Add a Puppetfile to track vendored modules
The Puppetfile describes the version of modules currently present in our
repo. There is one change to concat as it seems to not contain,
f507466942dbdb0684a1d04ea3d96d62d0ec70fa, though this was included in
v7.0.1, I am not sure why it is missing, but it seems like a cosmetic
change. All the other additional files are testing and support files.
They could be removed, but it would be nice if we could use tooling like
r10k & g10k to update modules, even if we do that infrequently.

Bug: T325597
Hosts: auto
Change-Id: I854454a9b1d9b3f139e2e0135594e364257b0028