phabricator: rewrite https://phabricator.wikimedia.org/r/ to gerrit
Adds a new rewrite conditition to check if the ServerName is
the main ServerName, not one of the alias.

Strips the protocol and replaces . with \. for proper use in the
RewriteCond.

Then adds a new rewrite rule to rewrite /r/, and only exactly /r/,
to https://gerrit.wikimedia.org/g/.

L flag to stop processing rules and G flag to send 410 Gone along
with it to tell the client this URL does not exit anymore.

Bug: T324311
Change-Id: I62e686915a18b66b559c209913341bbb5c6351a3