swift: add swift::rclone
Replace swiftrepl (to be removed in a separate CR) with a script based
on rclone.

Run this on a storage node ('rclone_host' in Swift::Cluster_info ),
since rclone uses relatively large amounts of RAM when constructing
container lists.

The set of containers to sync is based on what swifrepl was doing, but
removing the thumbs containers since they are very large indeed (and
we don't in fact care if thumbs get out of sync as they are
essentially just a cache anyway).

Use confctl to check which is the active DC, and only run the sync
from the active end. Like swiftrepl, this isn't very graceful if DC
switchover happens while it's running. Likewise, run just once a week,
on a Monday morning (swiftrepl ran on Monday from one DC and Wednesday
from the other).

Since the rclone nodes are storage hosts, we need
profile::swift::cluster_label available from both storage and proxy
nodes, so move it in hiera.

Bug: T299125
Hosts: auto
Change-Id: I35ed37a46dcd622ebdfe157e68c793fad628e8d5