jenkins: set systemd service start timeout to 5 min
When restart Jenkins this morning, the host was being backed up which
slow down disk I/O. As Jenkins is doing some migration and garbage
collection tasks on startup, it might took more than 90 seconds before
it is able to notify systemd it actually started. We went to disable the
systemd timeout to let it start.

Given it only happens when a backup is ongong and the migration steps
should not take much more than a couple minutes, change the timeout from
infinitiy to 5 minutes which should be enough time.

Change-Id: I54bd53f69b3b831f16fb8b52fe70221d8f3bb69f