httpd: add flag to wait for network-online.target
On Gerrit, Apache2 listens explicitly to the service address. We found
out after a reboot Apache fails to open a socket to the service IP
address cause it is not ready yet:

 systemd[1]: Starting The Apache HTTP Server...
 apachectl[846]: (99)Cannot assign requested address:
 AH00072: make_sock: could not bind to address [2620:0:861:2:208:80:154:137]:80
 apachectl[846]: no listening sockets available, shutting down
 apachectl[846]: AH00015: Unable to open logs
 apachectl[846]: Action 'start' failed.

The unit provided by Debian has `After=network.target` which only states
the network management stack is up but doesn't ensure the network
interfaces have been configured. One needs `network-online.target` for
that.

Add a boolean flag `wait_network_online` which when set to True would
inject a systemd override for the apache2 service with:

 [Unit]
 After=network-online.target
 Wants=network-online.target

Systemd would append the value the builtin lists. I have tested it
locally by creating a /etc/systemd/system/apache2.service.d/local.conf
and confirmed the new target is added to `After` and a `Wants` is
created.

Bug: T326125
Change-Id: I2c3d46f8dff3444a7cef3fec2e4387edeea214a2