ToolsDB: stop replicating a big problematic table
s52481__stats_global.requests caused the replica server clouddb1002 to
lag days behind of the primary, while trying to replicate a huge
transaction deleting millions of lines from that table.

Adding that table to the "Replicate_Wild_Ignore_Table" exclusion list
allowed the replication to resume. This is not ideal but we have already
other tables that were added to this list for similar reasons, and we
are gonna create a brand new replica server in T301949.

This change is already live on clouddb1002, but we need it to be in
Puppet as well, in case mariadb is restarted and the config is refreshed
from this file.

Bug: T326261
Change-Id: Ie7338dc3bd71382f0c940a9938b4a59d198b5fe4