Detect the correct disks for the O/S on the cephosd servers
Since it is no longer possible to rely on predictable device names for
block storage devices, we have to use another mechanism to ascertain
which devices should be used for the O/S when preseeding.

The new cephosd servers require this, so this commit adds a case for
their hostnames to the partman-early-command.sh file and a suitable
function to configure the debian-installer with the correct device names.

In this particular case we examine all of the block devices and select
only SSDs with a capacity of less than 3 TB. The two drives were are
using for the operating system are 480 GB SSDs.

Three RAID1 mirrors are created on these disks; for boot, swap, and an
LVM physical volume.

I have checked that the blockdev and wc and egrep commands referenced in
the function are available in the d-i environment for bullseye.

Bug: T324670
Change-Id: I1bc67686f7f9dc254629a932469b430696a9c83a