check_legal_terms: Refactor check to make it more robust against changes
Rather than using strict regular expressions (which anyway are not
suitable to parse HTML), change those into parsing HTML with the
BeautifulSoup library to detect the footer, and identify
keywords that must always be there, even if the spelling or order
changes.

Additionally, check that the link targets actually contain the
expected text, but containing key terms.

For example, we expect a link that contains the words "privacy"
and "policy", and we check that such a link (even if its url or
content changes in the future contains references to the words:

'wikimedia', 'introduction', 'site', 'personal information',
'third party', 'collect', 'contribution', 'metadata',
'information', 'sharing', 'protect' and 'contact'

ensuring that the text is still there.

This will allow communities to tune the text while also ensuring
that there are still references to:

* Copyright and a Creative Commons license
* Terms of use
* Privacy policy
* Wikipedia trademark

This way the mobile and desktop can reuse the same checks, even
if the mobile page contains a smaller version of the above.

Bug: T317169
Change-Id: Iec57ea7abb68b7543b9f5ecf0b1c43731466d154