admin/canary_appserver: add group of users allowed to disable puppet
As requested in the past, allow deployers to disable puppet
on mwdebug hosts for debugging.

But do so in a way that does not mean these privileges get applied
on all hosts deployers have access to, only on mwdebug hosts.

To achieve this make a new group mwdebuggers, give it the privileges
to run the enable-puppet and disable-puppet wrappers and copy the
members from the deployment group.

Finally put the group in the canary_appserver role and remove
the mediawiki-testers group since that is empty and has no members
and privileges it once had.

Bug: T305979
Change-Id: I8192b19a15d6aa006a4e67f7d1677fb093b85d6a