icinga:Update legal check to link to wikitech and add legal contact
The legal check terms now will have a runbook for both legal AND
systemd administrators at:

https://wikitech.wikimedia.org/wiki/Check_legal_html

Add it to the check definition so it can be available on icinga.

Also, add, in addition to slaporte, the legal ticketing system
so he is no longer a single point of failure, and more than one
person at legal can attend those alerts. This was ok-ed by legal.

Bug: T317169
Change-Id: I0ae77d14f03f8334f79e0aae9ec566a7f993a1a5