prometheus: Generate varnish params file
The thread pool settings have been templated out, so we can now export
the Puppet values to a node_exporter file. This will allow us to monitor
the threads_max values without a routine job (causing the varnishd
process to spew out all parameters to the syslog).

Note that thread_pool_max is multiplied by a hard-coded "2" because
Varnish has defaulted the thread_pools value to 2 and will be unlikely
to change that anytime soon. We want to continue using the default value
in case Varnish changes this in the future for better performance
reasons and risk the Prometheus data rather than have the more accurate
data but running a sub-optimal performance setting.

Bug: T323723
Change-Id: Ief5f55d0a60cfd0db85d760acab2f8f3cbd61804