Alerts: stop alerting on thumb number mismatch
We stopped copying thumbnails to the other DC in wmf/rewrite.py a
while back (while resolving T313102), and we don't worry about
replicating them later either - if necessary we can just make a new
one.

Given all of which, stop alerting if the number of thumbs in the two
DCs gets out of sync.

Bug: T326857
Hosts: alert1001.wikimedia.org
Change-Id: I86f150b195141744d723d50e845bac5bd915604f