peopleweb: ensure rsync service is stopped on passive host
The 'rsync' package is installed on both hosts but only one
of them will include the rsync::server class which sets it up
as service.

Ensure that service is always stopped on the passive host so
that even if we switch between servers we don't end up with a
host that is still running rsync without having the auto-restart
service for it.

Unless we want it to also purge /etc/rsync.d/ entirely...

follow-up to I2fcc47a1fb72f643cdc

Bug: T326888
Change-Id: I42f71f587d0546b495d17bda6a2991f01150c1f8