mediawiki: Add auto_prepend_file to PHP config_cli
This way when using mwscript on mwdebug or mwmaint hosts, options
like `--profiler` will work accordingly as right now a portion of MW
config is essentially missing.

Test plan:

* Today, the following print no stats:
  mwdebug1001$ mwscript showJobs.php --wiki testwiki --profiler text
  # which effectively expands to:
  mwdebug1001$ sudo -u www-data php /srv/mediawiki/multiversion/MWScript.php \
  showJobs.php --wiki testwiki --profiler text

* Today, the following workaround works:
  $ sudo -u www-data \
   php -d auto_prepend_file=/srv/mediawiki/wmf-config/PhpAutoPrepend.php \
   /srv/mediawiki/multiversion/MWScript.php showJobs.php \
   --wiki testwiki --profiler text

* With this change applied, the proper command should start printing
  stats, like the workaround does:
  mwdebug1001$ mwscript showJobs.php --wiki testwiki --profiler text

PhpAutoPrepend has been changed and audited to no-op on CLI the same
way it has always no-opped on web requests when profiler options aren't
in use. With regards to the Excimer sampling profiler, this is explicitly
disabled in wmf-config/src/Profiler.php for CLI and will remain limited
to web requests.

Bug: T253547
Change-Id: I135fcd22d3e9c82b1200210082179b505744e152