thumbor: add and use haproxy healthz lvs check
Currently we check /healthcheck via LVS for thumbor. If the check hits
a busy thumbor worker or if all workers are busy, the host will be
unhealthy. This has mostly been fine so far because we have too few
thumbor hosts pooled for lvs to depool unhealthy hosts too often.

This means that we are also not using haproxy's queuing behaviour
despite it being an expectation of our setup. It also means that
adding more Kubernetes hosts to pybal potentially puts us at risk of
depooling other hosts due to pybal seeing more "spare" hosts. It is
also a blocker for adding Kubernetes nodes given the smaller number of
workers.

The healthcheck will fail if the average queue length across all
workers exceeds a limit

Bug: T233196
Change-Id: I833553bedbe160acd812ccff67edf11c90848b2f